move assetProviding in AppDelegate
