+++ Example/Snabble/AppDelegate.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+import SnabbleSDK
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -13,6 +14,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+
+        Asset.provider = self
 
         let shopsNavi = UINavigationController(rootViewController: ShopsViewController())
         let accountNavi = UINavigationController(rootViewController: AccountViewController())
